MapBox API basic integration

+++ Assets/PlaceHardcodedWaypoint.cs
@@ -5,11 +5,11 @@ using UnityEngine;
 
 public class PlaceHardcodedWaypoint : MonoBehaviour {
     
-    public float waypointLatitude;
-    public float waypointLongitude;
+    public double waypointLatitude;
+    public double waypointLongitude;
     public float waypointAltitude;
 
-    public float initialLat;
+    public double initialLat;
 
     public GameObject hardcodedWaypointSphere;
 	// Use this for initialization
@@ -22,9 +22,9 @@ public class PlaceHardcodedWaypoint : MonoBehaviour {
         if (WorldProperties.droneInitialPositionSet)
         {
            Vector3 changePos = new Vector3(
-                   ((WorldProperties.LatDiffMeters(WorldProperties.droneHomeLat, waypointLatitude)) / WorldProperties.Unity_X_To_Lat_Scale),
-                   ((waypointAltitude /*- WorldProperties.droneHomeAlt*/) / WorldProperties.Unity_Y_To_Alt_Scale),
-                   ((WorldProperties.LongDiffMeters(WorldProperties.droneHomeLong, waypointLongitude, waypointLatitude) / WorldProperties.Unity_Z_To_Long_Scale))
+                   ((float) (WorldProperties.LatDiffMeters(WorldProperties.droneHomeLat, waypointLatitude)) / WorldProperties.Unity_X_To_Lat_Scale),
+                   ((float) (waypointAltitude /*- WorldProperties.droneHomeAlt*/) / WorldProperties.Unity_Y_To_Alt_Scale),
+                   ((float) (WorldProperties.LongDiffMeters(WorldProperties.droneHomeLong, waypointLongitude, waypointLatitude) / WorldProperties.Unity_Z_To_Long_Scale))
                  );
             initialLat = WorldProperties.droneHomeLat;
             this.transform.localPosition = changePos;
